Cleaned up fitures. Extracted static calcualtor into single class.

+++ RpnCalculator/src/com/rpn/RpnCalculator.java
@@ -19,7 +19,7 @@ public class RpnCalculator {
         this.factory = factory;
     }
 
-    public void enter(int value) {
+    public void enter(BigDecimal value) {
         values.push(value);
     }
